feat: split empty msg and address selection
There is actually an empty message right before the address selection where we originally thought this was "A prefix that does not count towards the checksum". Correct this by sending an empty IBIS message first and then the address selection. Get rid of the prefix complexity in the builder so that we do not need to worry about this.